style: rename column module to columns
